[SofaAllCommonComponents] RENAME to SofaComponentAll
This is more in phase with existing packages like SofaComponentAdvanced, etc.
