+++ SofaKernel/modules/sofa/frameworkextra/frameworkextra_test/core/objectmodel/BaseContext_test.cpp
@@ -39,7 +39,7 @@ public:
     void testGetObjects()
     {
         EXPECT_MSG_NOEMIT(Error, Warning) ;
-        importPlugin("SofaAllCommonComponents") ;
+        importPlugin("SofaComponentAll") ;
         std::stringstream scene ;
         scene << "<?xml version='1.0'?>"
                  "<Node name='Root' gravity='0 -9.81 0' time='0' animate='0' >               \n"
