+++ SofaKernel/modules/sofa/frameworkextra/frameworkextra_test/simulation/Node_test.cpp
@@ -43,7 +43,7 @@ public:
 
     NodeContext_test()
     {
-        importPlugin("SofaAllCommonComponents") ;
+        importPlugin("SofaComponentAll") ;
     }
 
     void testGetNodeObjects()
