+++ applications/packages/CMakeLists.txt
@@ -1,4 +1,4 @@
 cmake_minimum_required(VERSION 3.1)
 
-sofa_add_plugin(SofaAllCommonComponents SofaAllCommonComponents ON)
+sofa_add_plugin(SofaComponentAll SofaComponentAll ON)
