+++ applications/packages/SofaComponentAll/config.h
@@ -23,9 +23,9 @@
 #define SOFA_PACKAGE_ALLCOMMONCOMPONENTS_CONFIG_H
 
 #ifdef SOFA_BUILD_MISC_COLLISION
-#  define SOFA_SOFAALLCOMMONCOMPONENTS_API SOFA_EXPORT_DYNAMIC_LIBRARY
+#  define SOFA_SOFACOMPONENTALL_API SOFA_EXPORT_DYNAMIC_LIBRARY
 #else
-#  define SOFA_SOFAALLCOMMONCOMPONENTS_API SOFA_IMPORT_DYNAMIC_LIBRARY
+#  define SOFA_SOFACOMPONENTALL_API SOFA_IMPORT_DYNAMIC_LIBRARY
 #endif
 
 #endif
