+++ applications/packages/SofaComponentAll/initComponentAll.h
